Moved files for pages/components into the views folder
