+++ client/src/index.js
@@ -12,7 +12,7 @@ import "assets/scss/paper-dashboard.scss?v=1.1.0";
 import "assets/demo/demo.css";
 import "perfect-scrollbar/css/perfect-scrollbar.css";
 
-import NewSession from "./components/NewSession/NewSession"
+import NewSession from "./views/pages/forms/NewSession"
 
 const hist = createBrowserHistory();
