+++ client/src/routes.js
@@ -1,12 +1,12 @@
 import Calendar from "views/Calendar.jsx";
-import Login from "views/pages/Login.jsx";
-import Register from "views/pages/Register.jsx";
+import Login from "views/pages/forms/Login.jsx";
+import Register from "views/pages/forms/Register.jsx";
 import UserProfile from "views/pages/UserProfile.jsx";
-import ChildPage from "components/ChildPage/ChildPage";
-import OneResource from "components/OneResource/OneResource"
+import ChildPage from "views/pages/ChildPage";
+import OneResource from "views/pages/OneResource"
 import Resources from "views/pages/Resources";
-import AddChild from "./components/AddChild/AddChild";
-import NewSession from "./components/NewSession/NewSession";
+import AddChild from "./views/pages/forms/AddChild";
+import NewSession from "./views/pages/forms/NewSession";
 import ChildTable from "./views/tables/ChildTable";
 import Therapists from "views/pages/Therapists";
 import ViewSession from "views/pages/ViewSession";
