+++ client/src/views/pages/OneResource.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import withAuth from '../withAuth';
+import withAuth from '../../components/withAuth';
 import { Link } from 'react-router-dom';
 import {
   Card,
