+++ client/src/views/pages/forms/AddChild.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import withAuth from '../withAuth';
-import API from '../../utils/API';
+import withAuth from '../../../components/withAuth';
+import API from '../../../utils/API';
 import ReactDatetime from "react-datetime";
 import {
   Button,
